Move TcpTransportLayer to node
